% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_template}
\alias{data_template}
\title{data_template}
\format{
dataset with 32 rows and 5 columns:
\describe{
\item{observed}{Observed values}
\item{Reduced Model 1}{Predicted values for the Reduced Model 1.}
\item{Reduced Model 2}{Predicted values for the Reduced Model 2.}
\item{Reduced Model 3}{Predicted values for the Reduced Model 3.}
\item{Full Model}{Predicted values for the Full Model.}
...
}
}
\source{
Data simulated for this application.
}
\usage{
data_template
}
\description{
Template to be used for the user.
}
\keyword{datasets}
