% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{model_eval}
\alias{model_eval}
\title{model_eval}
\usage{
model_eval(obs_values, pred_values)
}
\arguments{
\item{obs_values}{observed values}

\item{pred_values}{predicted values}
}
\value{
list with the main statistics
}
\description{
model_eval
}
