% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval_ui.R
\name{model_eval_ui}
\alias{model_eval_ui}
\title{Graphical Interface for Model Evaluation}
\usage{
model_eval_ui()
}
\value{
graphical interface where the user must upload a .XLSX file to perform
the model evaluation. There is no parameters for this function.
}
\description{
Graphical Interface for Model Evaluation
}
