% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelEvalUI.R
\name{run_modelEvalUI}
\alias{run_modelEvalUI}
\title{Graphical Interface for Model Evaluation}
\usage{
run_modelEvalUI()
}
\value{
graphical interface where the user can upload a .XLSX file to perform
the model evaluation
}
\description{
Graphical Interface for Model Evaluation
}
