% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_template}
\alias{data_template}
\title{data_template}
\format{
Uma base com 8026 linhas e 8 colunas:
\describe{
\item{temporada}{Descrição da coluna 1}
\item{time}{Descrição da coluna 2}
\item{n_vitorias}{Descrição da coluna 2}
...
}
}
\source{
Origem dos dados \url{https://www.cbf.com.br/futebol-brasileiro/competicoes/campeonato-brasileiro-serie-a}
}
\usage{
data_template
}
\description{
Uma base que contém...
}
\keyword{datasets}
