% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model_eval}
\alias{model_eval}
\title{Performs model evaluation.}
\usage{
model_eval(obs_values, pred_values)
}
\arguments{
\item{obs_values}{observed values.}

\item{pred_values}{predicted values.}
}
\value{
Data frame with the statistical metrics to perform model evaluation and comparison.
}
\description{
Performs model evaluation by calculating several metrics available in the literature.
}
