% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred_obs.R
\name{plot_pred_obs}
\alias{plot_pred_obs}
\title{plot_pred_obs}
\usage{
plot_pred_obs(dataset)
}
\arguments{
\item{dataset}{provide a dataset to...}
}
\value{
a plot with ...
}
\description{
plot_pred_obs
}
